customize: Don't fail if a package is already gone
The makers of Ubuntu have started to cut some dependencies in order to not pull
into Ubuntu in system containers packages that are not needed there. Our script
'customize' will not fail anymore if a package has not been there in the first
place.

Charmap UTF-8 now depends on ANSI*, so we retain the latter as well.